moved non modules out of modules
