+++ InfernalRobotics/InfernalRobotics/Command/Translator.cs
@@ -1,25 +1,22 @@
 ﻿using System;
 using InfernalRobotics.Module;
 
-namespace InfernalRobotics
+namespace InfernalRobotics.Command
 {
-    /* 
+    /*
      * <summary>
-     * This class acts as a translator for UI - it implement basic commands for servos 
+     * This class acts as a translator for UI - it implement basic commands for servos
      * and translates them to internal routines for servo controller.
-     * 
-     * Thus later we can add more sophsticated controller commands, or change their behaviour 
+     *
+     * Thus later we can add more sophisticated controller commands, or change their behavior
      * per part basis without breaking the way UI works.
-     * 
+     *
      * More basic commands may be added too. Future API calls will most likely be hooked to this class.
-     * 
+     *
      */
+
     public class Translator
     {
-        public Translator()
-        {
-        }
-
         public void Init(Interpolator interpolator, float speedUnit)
         {
             Interpolator = interpolator;
@@ -34,22 +31,25 @@ namespace InfernalRobotics
         {
             Interpolator.SetCommand(pos, speed * SpeedUnit);
         }
+
         public void MoveIncremental(float pos, float speed)
         {
             Interpolator.SetIncrementalCommand(pos, speed * SpeedUnit);
         }
+
         public void Stop()
         {
             Move(0, 0);
         }
+
         public bool IsMoving()
         {
-            return Interpolator.Active && Interpolator.CmdVelocity != 0;
+            return Interpolator.Active && Math.Abs(Interpolator.CmdVelocity) > 0.0001;
         }
-        internal float getSpeedUnit()
+
+        internal float GetSpeedUnit()
         {
             return SpeedUnit;
         }
     }
 }
-
