+++ InfernalRobotics/InfernalRobotics/InfernalRobotics.csproj
@@ -53,6 +53,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Command\Interpolator.cs" />
+    <Compile Include="Command\Translator.cs" />
     <Compile Include="Effects\SoundSource.cs" />
     <Compile Include="Extension\ConfigNodeExtension.cs" />
     <Compile Include="Extension\RectExtensions.cs" />
@@ -60,10 +62,8 @@
     <Compile Include="Gui\ControlsGUI.cs" />
     <Compile Include="Gui\GUIDragAndDrop.cs" />
     <Compile Include="Gui\ToolbarWrapper.cs" />
-    <Compile Include="Module\Interpolator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Module\MuMechToggle.cs" />
-    <Compile Include="Module\Translator.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Image\counterclockwise.png" />
